+++ CHANGELOG.md
@@ -17,6 +17,10 @@
 - favicon (nginx setting? '/' restriction? )
 - css
 
+## 0.9.1 -2022-09-02
+- refactored continued. lastupdate, startdate, enddate.
+- refactored output
+
 ## 0.9.0 - 2022-09-02
 - refactor. use session globally in `core.cljs`.